This merge removes motion blur, for now. I used perliedman/raytracing-in-one-weekend
as a base for comparison and that repo does not include motion blur. So to make
comparisons more accurate I removed motion blur as well. I am planning on adding
it back in in the future.

Running with single threaded mode we perform close to perliedman's version (Which
is single threaded only). In multithreaded mode we run much much faster.

+++ src/utils.rs
@@ -3,7 +3,7 @@ use super::vec3::Vec3;
 
 pub fn random_in_unit_sphere() -> Vec3 {
     let mut rng = thread_rng();
-    let mut rnd = || rng.gen::<f64>();
+    let mut rnd = || rng.gen::<f32>();
     let mut sample = || 2.0 * Vec3::new(rnd(), rnd(), rnd()) - Vec3::new(1.0, 1.0, 1.0);
     let mut p = sample();
     while p.squared_length() >= 1.0 { p = sample(); }
